moving entities up.
